@@ -20,33 +20,31 @@
 	},
 	"dependencies": {
 		"@castlenine/svelte-qrcode": "^2.3.0",
-		"@typhonjs-fvtt/runtime": "^0.3.0-next.2",
-		"@typhonjs-fvtt/standard": "^0.3.0-next.2",
-		"nanoid": "^5.1.5",
-		"oauth4webapi": "^3.5.3",
+		"nanoid": "^5.1.6",
+		"oauth4webapi": "^3.8.2",
 		"postcss": "^8.5.6",
-		"sass": "^1.89.2",
-		"svelte": "^4.2.20"
+		"sass": "^1.93.2",
+		"svelte": "^5.39.8"
 	},
 	"devDependencies": {
-		"@antfu/eslint-config": "^4.16.1",
+		"@antfu/eslint-config": "^5.4.1",
 		"@svelte-plugins/tooltips": "^3.0.3",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/vite-plugin-svelte": "^6.2.1",
 		"autoprefixer": "^10.4.21",
-		"cssnano": "^7.0.7",
-		"esbuild": "^0.25.5",
-		"eslint": "^9.29.0",
+		"cssnano": "^7.1.1",
+		"esbuild": "^0.25.10",
+		"eslint": "^9.37.0",
 		"eslint-plugin-antfu": "^3.1.1",
-		"eslint-plugin-format": "^1.0.1",
-		"eslint-plugin-svelte": "^2.46.1",
-		"fvtt-types": "github:League-of-Foundry-Developers/foundry-vtt-types#commit=c25ecedb344909dcaf781ea5c47d40c8b6e27432",
-		"postcss-preset-env": "^10.2.3",
-		"rollup-plugin-visualizer": "^6.0.3",
-		"svelte-eslint-parser": "^0.43.0",
+		"eslint-plugin-format": "^1.0.2",
+		"eslint-plugin-svelte": "^3.12.4",
+		"fvtt-types": "13.346.0-beta.20251002185327",
+		"postcss-preset-env": "^10.4.0",
+		"rollup-plugin-visualizer": "^6.0.4",
+		"svelte-eslint-parser": "^1.3.3",
 		"svelte-preprocess": "^6.0.3",
-		"typescript": "^5.8.3",
-		"typescript-eslint": "^8.35.0",
-		"vite": "^6.3.5",
+		"typescript": "^5.9.3",
+		"typescript-eslint": "^8.45.0",
+		"vite": "^7.1.9",
 		"vitest": "^3.2.4"
 	},
 	"browserslist": [
